docs: 데이터베이스 성능 개선 보고서 개선 (기능 추가) refactor: 불필요 주석 제거, 코드 리뷰 반영 (OrderUseCase의 책임 분리)
