+++ src/test/java/kr/hhplus/be/server/order/application/service/unit/ProductServiceUnitTest.java
@@ -5,7 +5,7 @@ import kr.hhplus.be.server.common.exception.OutOfStockException;
 import kr.hhplus.be.server.common.exception.OutOfStockListException;
 import kr.hhplus.be.server.common.exception.ProductNotFoundException;
 import kr.hhplus.be.server.order.application.dto.OrderRequest;
-import kr.hhplus.be.server.order.application.service.ProductService;
+import kr.hhplus.be.server.order.application.domainService.ProductService;
 import kr.hhplus.be.server.product.domain.entity.Product;
 import kr.hhplus.be.server.product.infra.repository.port.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -97,7 +97,7 @@ class ProductServiceUnitTest {
         assertThat(result).hasSize(productList.size());
         assertThat(result.get(productList.get(0))).isEqualTo(5);
         assertThat(result.get(productList.get(1))).isEqualTo(3);
-      //  assertThat(result.get(productList.get(2))).isEqualTo(2);
+        assertThat(result.get(productList.get(2))).isEqualTo(2);
         Product firstKey = result.keySet().iterator().next();
         assertThat(firstKey.getName()).isEqualTo(productList.get(0).getName());
