+++ src/test/java/kr/hhplus/be/server/order/application/service/unit/UserServiceUnitTest.java
@@ -3,7 +3,7 @@ package kr.hhplus.be.server.order.application.service.unit;
 import kr.hhplus.be.server.common.exception.ErrorCode;
 import kr.hhplus.be.server.common.exception.InvalidRequestException;
 import kr.hhplus.be.server.common.exception.UserNotFoundException;
-import kr.hhplus.be.server.order.application.service.UserService;
+import kr.hhplus.be.server.order.application.domainService.UserService;
 import kr.hhplus.be.server.user.domain.entity.User;
 import kr.hhplus.be.server.user.infra.reposistory.port.UserRepository;
 import org.junit.jupiter.api.DisplayName;
